+++ TimeSpeed/Framework/LocationType.cs
@@ -9,7 +9,13 @@ namespace TimeSpeed.Framework
         /// <summary>The location is outside.</summary>
         Outdoors,
 
-        /// <summary>The mines or skull cavern.</summary>
-        Mine
+        /// <summary>The normal mines.</summary>
+        Mine,
+
+        /// <summary>The skull cavern.</summary>
+        SkullCavern,
+
+        /// <summary>The volcano dungeon.</summary>
+        VolcanoDungeon
     }
 }